Implemented chat mocks
