+++ byoeb-v1/byoeb/byoeb/services/chat/utils.py
@@ -79,33 +79,4 @@ def get_read_receipt_byoeb_messages(byoeb_messages: List[ByoebMessageContext]):
         if byoeb_message.message_category == MessageCategory.READ_RECEIPT.value
     ]
     return read_receipt_messages
-
-def get_mock_whatsapp_response(
-    phone_number_id: str,
-):
-    import uuid
-    from byoeb_core.models.whatsapp.response.message_response import (
-        WhatsAppResponse,
-        WhatsAppResponseStatus,
-        Contact,
-        Message
-    )
-    return WhatsAppResponse(
-        messaging_product="whatsapp",
-        response_status=WhatsAppResponseStatus(
-            status='200'
-        ),
-        contacts=[
-            Contact(
-                input=phone_number_id,
-                wa_id=phone_number_id
-            )
-        ],
-        messages=[
-            Message(
-                id=str(uuid.uuid4()),
-            )
-        ],
-        media_message=None
-    )
